fix python cylic imports and fail on CI if any are introduced
