+++ pyphare/pyphare/pharesee/plotting.py
@@ -66,9 +66,7 @@ def dist_plot(particles, **kwargs):
             raise ValueError("Only velocity Y-axis are supported yet")
 
         bins = kwargs.get("bins", (50, 50))
-        h, xh, yh = np.histogram2d(
-            x, y, bins=bins, weights=particles.weights[:, 0]
-        )
+        h, xh, yh = np.histogram2d(x, y, bins=bins, weights=particles.weights[:, 0])
 
         if "gaussian_filter_sigma" in kwargs and "median_filter_size" not in kwargs:
             from scipy.ndimage import gaussian_filter
@@ -80,7 +78,9 @@ def dist_plot(particles, **kwargs):
 
             siz = kwargs.get("median_filter_size", (0, 0))
             image = median_filter(h.T, size=siz)
-        elif "gaussian_filter_sigma" not in kwargs and "median_filter_size" not in kwargs:
+        elif (
+            "gaussian_filter_sigma" not in kwargs and "median_filter_size" not in kwargs
+        ):
             image = h.T
         else:
             raise ValueError(
@@ -102,7 +102,9 @@ def dist_plot(particles, **kwargs):
             elif color_scale == "linear":
                 norm = Normalize(cmin, cmax)
             else:
-                raise ValueError("Only log and linear color_scale values are supported yet")
+                raise ValueError(
+                    "Only log and linear color_scale values are supported yet"
+                )
 
             im = ax.pcolormesh(xh, yh, image, cmap=cmap, norm=norm)
 
@@ -114,14 +116,16 @@ def dist_plot(particles, **kwargs):
                 raise ValueError("stride must be a positive integer")
             markersize = kwargs.get("markersize", 0.5)
             alpha = kwargs.get("alpha", 0.5)
-            im = ax.scatter(x[::stride], y[::stride],
-                            color=color,
-                            marker='.',
-                            markersize=markersize,
-                            alpha=alpha)
+            im = ax.scatter(
+                x[::stride],
+                y[::stride],
+                color=color,
+                marker=".",
+                markersize=markersize,
+                alpha=alpha,
+            )
 
         ax.set_ylabel(kwargs.get("ylabel", axis[1]))
-
 
     elif len(axis) == 1:
         bins = kwargs.get("bins", (50))
@@ -130,6 +134,7 @@ def dist_plot(particles, **kwargs):
 
         if cuts is not None:
             from pyphare.core.box import Box
+
             if ndim == 1:
                 if cuts[0] is None:
                     new_particles = particles
@@ -137,10 +142,15 @@ def dist_plot(particles, **kwargs):
                     box_new = Box(cuts[0][0], cuts[0][1])
                     new_particles = particles.select(box_new, box_type="pos")
             elif ndim == 2:
-                box_new = Box((cuts[0][0], cuts[1][0]), (cuts[0][1], cuts[1][1]))  # TODO need to be tested
+                box_new = Box(
+                    (cuts[0][0], cuts[1][0]), (cuts[0][1], cuts[1][1])
+                )  # TODO need to be tested
                 new_particles = particles.select(box_new, box_type="pos")
             else:
-                box_new = Box((cuts[0][0], cuts[1][0], cuts[2][0]), (cuts[0][1], cuts[1][1], cuts[2][1]))  # TODO need to be tested
+                box_new = Box(
+                    (cuts[0][0], cuts[1][0], cuts[2][0]),
+                    (cuts[0][1], cuts[1][1], cuts[2][1]),
+                )  # TODO need to be tested
                 new_particles = particles.select(box_new, box_type="pos")
         else:
             new_particles = particles
@@ -150,20 +160,21 @@ def dist_plot(particles, **kwargs):
         if axis[0] in vaxis:
             x = new_particles.v[:, vaxis[axis[0]]]
         else:
-            raise ValueError("For 1-D dist_plot, the abscissa has to be a velocity axis")
+            raise ValueError(
+                "For 1-D dist_plot, the abscissa has to be a velocity axis"
+            )
 
         bins = kwargs.get("bins", 50)
-        h, xh = np.histogram(
-            x, bins=bins, weights=new_particles.weights[:, 0]
-        )
+        h, xh = np.histogram(x, bins=bins, weights=new_particles.weights[:, 0])
 
-        xh_ = 0.5*(xh[:-1]+xh[1:])
+        xh_ = 0.5 * (xh[:-1] + xh[1:])
         im = ax.plot(xh_, h, drawstyle=drawstyle)
 
         ax.set_ylabel(kwargs.get("ylabel", "f"))
 
     if kwargs.get("kde", False) is True:
         import seaborn as sns
+
         if len(axis) == 1:
             sns.kdeplot(x=x, ax=ax, color="w")
         else:
@@ -182,7 +193,8 @@ def dist_plot(particles, **kwargs):
             if axis[0] in vaxis:
                 ax.axvline(
                     np.average(
-                        new_particles.v[:, vaxis[axis[0]]], weights=new_particles.weights
+                        new_particles.v[:, vaxis[axis[0]]],
+                        weights=new_particles.weights,
                     ),
                     color="w",
                     ls="--",
@@ -190,7 +202,8 @@ def dist_plot(particles, **kwargs):
             if axis[1] in vaxis:
                 ax.axhline(
                     np.average(
-                        new_particles.v[:, vaxis[axis[1]]], weights=new_particles.weights
+                        new_particles.v[:, vaxis[axis[1]]],
+                        weights=new_particles.weights,
                     ),
                     color="w",
                     ls="--",
@@ -288,124 +301,3 @@ def zoom_effect(ax1, ax2, xmin, xmax, **kwargs):
     ax2.add_patch(p)
 
     return c1, c2, bbox_patch1, bbox_patch2, p
-
-
-def finest_field_plot(run_path, qty, **kwargs):
-    """
-    plot the given quantity (qty) at 'run_path' with only the finest data
-
-    * run_path : the path of the run
-    * qty : ['Bx', 'By', 'Bz', 'Ex', 'Ey', 'Ez', 'Fx', 'Fy', 'Fz', 'Vx', 'Vy', 'Vz', 'rho']
-
-    kwargs:
-    * ax : the handle for the fig axes
-    * time : time (that should be in the time_stamps of the run)
-    * interp : the type of interpolation for the interpolator
-    * draw_style : steps-mid ou default... only for 1d
-    * title : (str) title of the plot
-    * xlabel, ylabel
-    * xlim, ylim
-    * filename : (str) if exists, save plot to figure under that name
-
-    return value : fig,ax
-    """
-
-    import os
-    from pyphare.pharesee.hierarchy.fromh5 import get_times_from_h5
-    from pyphare.pharesee.run import Run
-    from mpl_toolkits.axes_grid1 import make_axes_locatable
-
-    r = Run(run_path)
-
-    time = kwargs.get("time", None)
-    dim = r.GetDl("finest", time).shape[0]
-    interp = kwargs.get("interp", "nearest")
-
-    if qty in ["Bx", "By", "Bz"]:
-        file = os.path.join(run_path, "EM_B.h5")
-        if time is None:
-            times = get_times_from_h5(file)
-            time = times[0]
-        interpolator, finest_coords = r.GetB(time, merged=True, interp=interp)[qty]
-    elif qty in ["Ex", "Ey", "Ez"]:
-        file = os.path.join(run_path, "EM_E.h5")
-        if time is None:
-            times = get_times_from_h5(file)
-            time = times[0]
-        interpolator, finest_coords = r.GetE(time, merged=True, interp=interp)[qty]
-    elif qty in ["Vx", "Vy", "Vz"]:
-        file = os.path.join(run_path, "ions_bulkVelocity.h5")
-        if time is None:
-            times = get_times_from_h5(file)
-            time = times[0]
-        interpolator, finest_coords = r.GetVi(time, merged=True, interp=interp)[qty]
-    elif qty == "rho":
-        file = os.path.join(run_path, "ions_charge_density.h5")
-        if time is None:
-            times = get_times_from_h5(file)
-            time = times[0]
-        interpolator, finest_coords = r.GetNi(time, merged=True, interp=interp)[qty]
-    elif qty in ("Jx", "Jy", "Jz"):
-        file = os.path.join(run_path, "EM_B.h5")
-        if time is None:
-            times = get_times_from_h5(file)
-            time = times[0]
-        interpolator, finest_coords = r.GetJ(time, merged=True, interp=interp)[qty]
-    else:
-        # ___ TODO : should also include the files for a given population
-        raise ValueError(
-            "qty should be in ['Bx', 'By', 'Bz', 'Ex', 'Ey', 'Ez', 'Fx', 'Fy', 'Fz', 'Vx', 'Vy', 'Vz', 'rho']"
-        )
-
-    if "ax" not in kwargs:
-        fig, ax = plt.subplots()
-    else:
-        ax = kwargs["ax"]
-        fig = ax.figure
-
-    if dim == 1:
-        drawstyle = kwargs.get("drawstyle", "steps-mid")
-
-        ax.plot(finest_coords[0], interpolator(finest_coords[0]), drawstyle=drawstyle)
-    elif dim == 2:
-        x = finest_coords[0]
-        y = finest_coords[1]
-        dx = x[1] - x[0]
-        dy = y[1] - y[0]
-
-        # pcolormesh considers DATA_ij to be the center of the pixel
-        # and X,Y are the corners so XY need to be made 1 value larger
-        # and shifted around DATA_ij
-        x -= dx / 2
-        x = np.append(x, x[-1] + dx)
-        y -= dy / 2
-        y = np.append(y, y[-1] + dy)
-
-        X, Y = np.meshgrid(x, y)
-        DATA = interpolator(X, Y)
-
-        vmin = kwargs.get("vmin", np.nanmin(DATA))
-        vmax = kwargs.get("vmax", np.nanmax(DATA))
-        cmap = kwargs.get("cmap", "Spectral_r")
-        im = ax.pcolormesh(x, y, DATA, cmap=cmap, vmin=vmin, vmax=vmax)
-        ax.set_aspect("equal")
-        divider = make_axes_locatable(ax)
-        cax = divider.append_axes("right", size="5%", pad=0.08)
-        plt.colorbar(im, ax=ax, cax=cax)
-    else:
-        raise ValueError("finest_field_plot not yet ready for 3d")
-
-    ax.set_title(kwargs.get("title", ""))
-
-    ax.set_xlabel(kwargs.get("xlabel", "$x \\ (d_p)$"))
-    ax.set_ylabel(kwargs.get("ylabel", "$y \\ (d_p)$"))
-
-    if "xlim" in kwargs:
-        ax.set_xlim(kwargs["xlim"])
-    if "ylim" in kwargs:
-        ax.set_ylim(kwargs["ylim"])
-
-    if "filename" in kwargs:
-        fig.savefig(kwargs["filename"])
-
-    return fig, ax
